backend/requirements.txt
Flask>=3.0.0
psycopg2-binary>=2.9.9    # PostgreSQL driver for Python
python-dotenv>=1.0.0
gunicorn>=22.0.0
